Separate ART renderer host config
